fix: convert int64 seed to uint64 for math/rand/v2.NewPCG
CodeRabbit flagged that rand.NewPCG requires uint64 args, but
resolveRandomSeed returns int64. The file is build-tagged e2e && stress
so go build ./... never compiled it; the error only surfaces under
the documented stress build.

Convert at the call site to keep resolveRandomSeed's int64 contract
and its 'not a valid int64' parse error intact. Bit pattern is
preserved so the same E2E_RANDOM_SEED value reproduces the run.

+++ e2e_stress_test.go
@@ -353,7 +353,7 @@ func TestMCPStress_Randomized(t *testing.T) {
 	seed := resolveRandomSeed(t)
 	t.Logf("randomized seed: %d (set E2E_RANDOM_SEED to replay)", seed)
 
-	rng := rand.New(rand.NewPCG(seed, seed))
+	rng := rand.New(rand.NewPCG(uint64(seed), uint64(seed)))
 	numSearches := 10
 	searchArgs := make([]map[string]any, numSearches)
 	searchQueries := make([]string, numSearches)
